Created STMulator
Stmulator is a class that does STM and didv simulations (and AFM
with ProbeParticleModel) of molecules. User has to create an instance
of the class with the desired parameters and then the instance can be
called with e.g.
    stmulator = STMulator()
    X = stmulator(xyz, Zs, qs, eigs, coefs)

Here, xyz is the atomic positions, Zs is the atomic numbers,
qs is the atomic charges, eigs is the eigenenergies and coefs
is the KS eigenvectors. The charges are point charges. In the future,
the class should be made to utilize hartree potentials instead.